Account CR backend, lost item CRUD, testing script/containers

+++ ss-services.sh
@@ -13,12 +13,15 @@ elif [[ "$1" = "test" ]]
 then
     cd webapp
     echo "Running webapp unit tests..."
-    python3 -m unittest discover -s test -p unit_*.py && \
+    python3 -m unittest discover -s test -p unit_*.py
 
-    cd .. && \
-    echo "Running webapp integration tests..." && \
-    docker compose -f docker-compose.test.yaml up --abort-on-container-exit --exit-code-from test-webapp
-    docker compose -f docker-compose.test.yaml down --volumes --remove-orphans
+    if [[ $? -eq 0 ]]
+    then
+        cd ..
+        echo "Running webapp integration tests..."
+        docker compose -f docker-compose.test.yaml up --abort-on-container-exit --exit-code-from test-webapp
+        docker compose -f docker-compose.test.yaml down --volumes --remove-orphans
+    fi
 else
     echo "ERROR: incorrect goal, must be [start|stop|restart|test]"
 fi
